Clean up file
